バックエンド: API安定化・テスト完全成功・型定義統一完了
## 主要改善
- 全86テスト完全成功 (100%パス率達成)
- repeat_intention値を英語統一 (yes/maybe/no)
- ランキングAPI新仕様完全対応 (shops配列形式)
- レート制限テスト修正 (ユーザーベース動作確認)

## API修正
- ReviewController: memo/comment型統一、詳細ログ追加
- RankingController: shops配列対応、バリデーション強化
- ShopController: バリデーション改善、エラーログ追加
- ImageController: サイズ別URL対応完了

## テスト修正
- RankingApiTest: 新API仕様対応
- RateLimitTest: キャッシュクリア処理追加
- ReviewApiTest: repeat_intention英語値対応
- ImageUploadTest: ファイルサイズ制限統一

## インフラ改善
- レート制限設定最適化 (開発環境緩和)
- 画像サービス URL生成改善
- JWT設定型安全性向上
- .htaccess アップロード制限追加

+++ backend/config/jwt.php
@@ -101,7 +101,7 @@ return [
     |
     */
 
-    'ttl' => env('JWT_TTL', 10080), // 1 week in minutes
+    'ttl' => (int) env('JWT_TTL', 10080), // 1 week in minutes
 
     /*
     |--------------------------------------------------------------------------
@@ -120,7 +120,7 @@ return [
     |
     */
 
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160),
 
     /*
     |--------------------------------------------------------------------------
